# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gojo API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/properties/": {
            "get": {
                "tags": [
                    "Manage Properties"
                ],
                "summary": "List available properties.",
                "operationId": "list-properties",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "API endpoint not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Manage Properties"
                ],
                "summary": "Create new property.",
                "operationId": "create-property",
                "parameters": [
                    {
                        "name": "property_title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "property_description",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "property_type_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "currency"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "decimal"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "decimal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "API endpoint not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "passport": {
            "type": "oauth2",
            "description": "Laravel passport oauth2 security.",
            "in": "header",
            "scheme": "https",
            "flows": {
                "password": {
                    "authorizationUrl": "http://gojo-api.test/oauth/authorize",
                    "tokenUrl": "http://gojo-api.test/oauth/token",
                    "refreshUrl": "http://gojo-api.test/token/refresh",
                    "scopes": []
                }
            }
        }
    }
}